Extract SyntaxTheme into its own crate
Move SyntaxTheme from theme/src/styles/syntax.rs into a new
syntax_theme crate that depends only on gpui. The theme crate
re-exports it for backward compatibility.
